refactor: split semantic proof by responsibility
Extract semantic-proof report models, fact extraction helpers, and surface-specific proof functions into focused modules while keeping the public semantic_proof API stable. Also align the dashboard TUI test with the current repository-only app contract.
