+++ polylogue/rendering/semantic_proof.py
@@ -3,19 +3,10 @@
 from __future__ import annotations
 
 import asyncio
-import csv
-import io
-import json
-import re
-from collections import Counter
-from dataclasses import dataclass, field
-from html import unescape as html_unescape
 from pathlib import Path
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 from polylogue.cli.query_output import format_summary_list, render_stream_transcript
-from polylogue.lib.models import ConversationSummary
-from polylogue.lib.roles import Role
 from polylogue.mcp.payloads import (
     MCPConversationDetailPayload,
     MCPConversationSummaryListPayload,
@@ -25,12 +16,31 @@ from polylogue.paths import archive_root as default_archive_root
 from polylogue.paths import db_path as default_db_path
 from polylogue.rendering.core import ConversationFormatter
 from polylogue.rendering.formatting import format_conversation
+from polylogue.rendering.semantic_proof_models import (
+    ProviderSemanticProof,
+    SemanticConversationProof,
+    SemanticMetricCheck,
+    SemanticProofReport,
+    SemanticProofSuiteReport,
+    _build_provider_reports,
+)
+from polylogue.rendering.semantic_proof_surfaces import (
+    _prove_mcp_detail_surface,
+    _prove_mcp_summary_surface,
+    _prove_query_stream_json_lines_surface,
+    _prove_query_stream_markdown_surface,
+    _prove_query_stream_plaintext_surface,
+    _prove_query_summary_csv_surface,
+    _prove_query_summary_json_like_surface,
+    _prove_query_summary_text_surface,
+    prove_export_surface_semantics,
+    prove_markdown_projection_semantics,
+)
 from polylogue.storage.backends.async_sqlite import SQLiteBackend
 from polylogue.storage.repository import ConversationRepository
-from polylogue.storage.store import ConversationRenderProjection, MessageRecord
 
 if TYPE_CHECKING:
-    from polylogue.lib.models import Conversation, Message
+    from polylogue.lib.models import Conversation
 
 
 DEFAULT_SEMANTIC_SURFACES: tuple[str, ...] = (
@@ -127,895 +137,6 @@ _EXPORT_SURFACE_FORMATS: dict[str, str] = {
     "export_org_v1": "org",
 }
 
-_HTML_ROLE_RE = re.compile(r'class="role-label">([^<]+)</span>')
-_HTML_TIMESTAMP_RE = re.compile(r'class="timestamp">[^<]+</time>')
-_HTML_BRANCH_RE = re.compile(r'class="branch-label">Branch\s+\d+</div>')
-_HTML_TITLE_RE = re.compile(r"<h1>(.*?)</h1>", re.DOTALL)
-_HTML_BADGE_RE = re.compile(r'class="badge">([^<]+)</span>')
-_HTML_META_RE = re.compile(r'class="meta-item">([^<]+)</span>')
-
-
-def _sorted_counts(counts: dict[str, int]) -> dict[str, int]:
-    return dict(sorted(counts.items()))
-
-
-def _empty_metric_counts() -> dict[str, int]:
-    return {"preserved": 0, "declared_loss": 0, "critical_loss": 0}
-
-
-def _normalized_role_label(value: object) -> str:
-    if isinstance(value, Role):
-        return str(value)
-    if value:
-        return str(Role.normalize(str(value)))
-    return "message"
-
-
-def _is_text_message(message: Message | MessageRecord) -> bool:
-    return bool((message.text or "").strip())
-
-
-def _critical_or_preserved(*, metric: str, policy: str, input_value: Any, output_value: Any) -> SemanticMetricCheck:
-    return SemanticMetricCheck(
-        metric=metric,
-        status="preserved" if input_value == output_value else "critical_loss",
-        policy=policy,
-        input_value=input_value,
-        output_value=output_value,
-    )
-
-
-def _declared_loss_or_preserved(*, metric: str, policy: str, input_value: int, output_value: Any = 0) -> SemanticMetricCheck:
-    return SemanticMetricCheck(
-        metric=metric,
-        status="declared_loss" if input_value else "preserved",
-        policy=policy,
-        input_value=input_value,
-        output_value=output_value,
-    )
-
-
-def _presence_check(
-    *,
-    metric: str,
-    policy: str,
-    input_value: object,
-    output_present: bool,
-) -> SemanticMetricCheck:
-    expected_present = input_value is not None
-    return SemanticMetricCheck(
-        metric=metric,
-        status="preserved" if expected_present == output_present else "critical_loss",
-        policy=policy,
-        input_value=input_value,
-        output_value=output_present,
-    )
-
-
-def _input_facts(projection: ConversationRenderProjection) -> dict[str, Any]:
-    attachment_counts: Counter[str] = Counter(
-        attachment.message_id for attachment in projection.attachments if attachment.message_id
-    )
-    renderable_messages = 0
-    timestamped_renderable_messages = 0
-    empty_messages = 0
-    thinking_messages = 0
-    tool_messages = 0
-    role_counts: Counter[str] = Counter()
-
-    for message in projection.messages:
-        has_attachments = attachment_counts.get(message.message_id, 0) > 0
-        has_text = _is_text_message(message)
-        if has_text or has_attachments:
-            renderable_messages += 1
-            role_counts[_normalized_role_label(message.role)] += 1
-            if message.sort_key is not None:
-                timestamped_renderable_messages += 1
-        else:
-            empty_messages += 1
-        if int(message.has_thinking or 0) > 0:
-            thinking_messages += 1
-        if int(message.has_tool_use or 0) > 0:
-            tool_messages += 1
-
-    return {
-        "total_messages": len(projection.messages),
-        "renderable_messages": renderable_messages,
-        "timestamped_renderable_messages": timestamped_renderable_messages,
-        "attachment_count": len(projection.attachments),
-        "empty_messages": empty_messages,
-        "thinking_messages": thinking_messages,
-        "tool_messages": tool_messages,
-        "renderable_role_counts": _sorted_counts(dict(role_counts)),
-    }
-
-
-def _canonical_markdown_output_facts(markdown_text: str) -> dict[str, Any]:
-    message_sections = 0
-    timestamp_lines = 0
-    attachment_lines = 0
-    role_section_counts: Counter[str] = Counter()
-
-    for line in markdown_text.splitlines():
-        if line.startswith("## "):
-            section = line[3:].strip()
-            if section.lower() != "attachments":
-                message_sections += 1
-                role_section_counts[_normalized_role_label(section)] += 1
-        elif line.startswith("_Timestamp: "):
-            timestamp_lines += 1
-        elif line.startswith("- Attachment: "):
-            attachment_lines += 1
-
-    return {
-        "message_sections": message_sections,
-        "timestamp_lines": timestamp_lines,
-        "attachment_lines": attachment_lines,
-        "typed_thinking_markers": 0,
-        "typed_tool_markers": 0,
-        "role_section_counts": _sorted_counts(dict(role_section_counts)),
-    }
-
-
-def _conversation_input_facts(conversation: Conversation) -> dict[str, Any]:
-    role_counts: Counter[str] = Counter()
-    message_ids: list[str] = []
-    text_message_ids: list[str] = []
-    timestamped_text_messages = 0
-    attachment_count = 0
-    thinking_messages = 0
-    tool_messages = 0
-    branch_messages = 0
-
-    for message in conversation.messages:
-        message_ids.append(str(message.id))
-        attachment_count += len(message.attachments)
-        if message.is_thinking:
-            thinking_messages += 1
-        if message.is_tool_use:
-            tool_messages += 1
-        if message.branch_index > 0:
-            branch_messages += 1
-        if _is_text_message(message):
-            text_message_ids.append(str(message.id))
-            role_counts[_normalized_role_label(message.role)] += 1
-            if message.timestamp is not None:
-                timestamped_text_messages += 1
-
-    display_date = conversation.display_date.isoformat() if conversation.display_date else None
-    return {
-        "conversation_id": str(conversation.id),
-        "provider": str(conversation.provider),
-        "title": conversation.display_title,
-        "date": display_date,
-        "total_messages": len(conversation.messages),
-        "text_messages": len(text_message_ids),
-        "message_ids": message_ids,
-        "text_message_ids": text_message_ids,
-        "text_role_counts": _sorted_counts(dict(role_counts)),
-        "timestamped_text_messages": timestamped_text_messages,
-        "attachment_count": attachment_count,
-        "thinking_messages": thinking_messages,
-        "tool_messages": tool_messages,
-        "branch_messages": branch_messages,
-    }
-
-
-def _summary_input_facts(summary: ConversationSummary, message_count: int) -> dict[str, Any]:
-    return {
-        "conversation_id": str(summary.id),
-        "provider": str(summary.provider),
-        "title": summary.display_title,
-        "date": summary.display_date.isoformat() if summary.display_date else None,
-        "messages": message_count,
-        "tags": list(summary.tags),
-        "summary": summary.summary,
-    }
-
-
-def _summary_output_facts(payload: dict[str, Any]) -> dict[str, Any]:
-    tags = payload.get("tags")
-    if not isinstance(tags, list):
-        tags = []
-    return {
-        "conversation_id": str(payload.get("id") or ""),
-        "provider": str(payload.get("provider") or ""),
-        "title": payload.get("title"),
-        "date": payload.get("date"),
-        "messages": int(payload.get("messages") or 0),
-        "tags": [str(tag) for tag in tags],
-        "summary": payload.get("summary"),
-    }
-
-
-def _summary_csv_output_facts(csv_text: str) -> list[dict[str, Any]]:
-    reader = csv.DictReader(io.StringIO(csv_text))
-    rows: list[dict[str, Any]] = []
-    for row in reader:
-        tags_text = str(row.get("tags") or "")
-        rows.append(
-            {
-                "conversation_id": str(row.get("id") or ""),
-                "provider": str(row.get("provider") or ""),
-                "title": row.get("title"),
-                "date": row.get("date") or None,
-                "messages": int(row.get("messages") or 0),
-                "tags": [tag for tag in tags_text.split(",") if tag],
-                "summary": row.get("summary") or None,
-            }
-        )
-    return rows
-
-
-_SUMMARY_TEXT_RE = re.compile(
-    r"^(?P<id>.{1,24})\s{2,}(?P<date>\S*)\s{2,}(?P<provider>\S+)\s{2,}(?P<title>.*) \((?P<messages>\d+) msgs\)$"
-)
-
-
-def _summary_text_output_facts(text: str) -> list[dict[str, Any]]:
-    rows: list[dict[str, Any]] = []
-    for line in text.splitlines():
-        match = _SUMMARY_TEXT_RE.match(line.rstrip())
-        if match is None:
-            continue
-        rows.append(
-            {
-                "conversation_id_prefix": match.group("id").strip(),
-                "provider": match.group("provider").strip(),
-                "date": match.group("date").strip() or None,
-                "title": match.group("title").strip(),
-                "messages": int(match.group("messages")),
-            }
-        )
-    return rows
-
-
-def _stream_input_facts(
-    conversation: Conversation,
-    *,
-    dialogue_only: bool = False,
-    message_limit: int | None = None,
-) -> dict[str, Any]:
-    filtered_messages = [
-        message
-        for message in conversation.messages
-        if (not dialogue_only or message.is_dialogue)
-    ]
-    if message_limit is not None:
-        filtered_messages = filtered_messages[:message_limit]
-
-    visible_messages = [message for message in filtered_messages if _is_text_message(message)]
-    role_counts: Counter[str] = Counter(_normalized_role_label(message.role) for message in visible_messages)
-    timestamped_messages = sum(1 for message in visible_messages if message.timestamp is not None)
-
-    return {
-        "conversation_id": str(conversation.id),
-        "provider": str(conversation.provider),
-        "title": conversation.display_title,
-        "date": conversation.display_date.isoformat() if conversation.display_date else None,
-        "text_messages": len(visible_messages),
-        "text_message_ids": [str(message.id) for message in visible_messages],
-        "text_role_counts": _sorted_counts(dict(role_counts)),
-        "timestamped_text_messages": timestamped_messages,
-        "attachment_count": sum(len(message.attachments) for message in filtered_messages),
-        "thinking_messages": sum(1 for message in filtered_messages if message.is_thinking),
-        "tool_messages": sum(1 for message in filtered_messages if message.is_tool_use),
-        "branch_messages": sum(1 for message in filtered_messages if message.branch_index > 0),
-        "dialogue_only": dialogue_only,
-        "message_limit": message_limit,
-    }
-
-
-def _stream_markdown_output_facts(text: str) -> dict[str, Any]:
-    title: str | None = None
-    provider: str | None = None
-    has_date = False
-    message_sections = 0
-    role_counts: Counter[str] = Counter()
-    for line in text.splitlines():
-        if title is None and line.startswith("# "):
-            title = line[2:].strip()
-        elif line.startswith("**Provider**: "):
-            provider = line.split(": ", 1)[1].strip()
-        elif line.startswith("**Date**: "):
-            has_date = True
-        elif line.startswith("## "):
-            role = line[3:].strip()
-            message_sections += 1
-            role_counts[_normalized_role_label(role)] += 1
-    footer_match = re.search(r"_Streamed (\d+) messages_", text)
-    return {
-        "title": title,
-        "provider": provider,
-        "has_date": has_date,
-        "message_sections": message_sections,
-        "role_counts": _sorted_counts(dict(role_counts)),
-        "footer_count": int(footer_match.group(1)) if footer_match else 0,
-    }
-
-
-def _stream_plaintext_output_facts(text: str) -> dict[str, Any]:
-    role_counts: Counter[str] = Counter()
-    message_sections = 0
-    for line in text.splitlines():
-        if line.startswith("[") and line.endswith("]"):
-            role = line[1:-1].strip()
-            message_sections += 1
-            role_counts[_normalized_role_label(role)] += 1
-    return {
-        "message_sections": message_sections,
-        "role_counts": _sorted_counts(dict(role_counts)),
-        "title": None,
-        "provider": None,
-        "has_date": False,
-    }
-
-
-def _stream_json_lines_output_facts(text: str) -> dict[str, Any]:
-    header: dict[str, Any] = {}
-    footer: dict[str, Any] = {}
-    message_ids: list[str] = []
-    role_counts: Counter[str] = Counter()
-    timestamped_messages = 0
-
-    for line in text.splitlines():
-        stripped = line.strip()
-        if not stripped:
-            continue
-        try:
-            payload = json.loads(stripped)
-        except json.JSONDecodeError:
-            continue
-        record_type = payload.get("type")
-        if record_type == "header":
-            header = payload
-        elif record_type == "footer":
-            footer = payload
-        elif record_type == "message":
-            message_ids.append(str(payload.get("id") or ""))
-            role_counts[_normalized_role_label(payload.get("role"))] += 1
-            if payload.get("timestamp"):
-                timestamped_messages += 1
-
-    return {
-        "conversation_id": str(header.get("conversation_id") or ""),
-        "title": header.get("title"),
-        "provider": str(header.get("provider") or ""),
-        "date": header.get("date"),
-        "message_ids": message_ids,
-        "message_sections": len(message_ids),
-        "role_counts": _sorted_counts(dict(role_counts)),
-        "timestamped_messages": timestamped_messages,
-        "footer_count": int(footer.get("message_count") or 0),
-    }
-
-
-def _mcp_summary_output_facts(payload: dict[str, Any]) -> dict[str, Any]:
-    return {
-        "conversation_id": str(payload.get("id") or ""),
-        "provider": str(payload.get("provider") or ""),
-        "title": payload.get("title"),
-        "messages": int(payload.get("message_count") or 0),
-        "created_at": payload.get("created_at"),
-        "updated_at": payload.get("updated_at"),
-    }
-
-
-def _mcp_detail_input_facts(conversation: Conversation) -> dict[str, Any]:
-    input_facts = _conversation_input_facts(conversation)
-    return {
-        "conversation_id": input_facts["conversation_id"],
-        "provider": input_facts["provider"],
-        "title": input_facts["title"],
-        "created_at": conversation.created_at.isoformat() if conversation.created_at else None,
-        "updated_at": conversation.updated_at.isoformat() if conversation.updated_at else None,
-        "messages": input_facts["total_messages"],
-        "message_ids": input_facts["message_ids"],
-        "role_counts": input_facts["text_role_counts"],
-        "timestamped_messages": input_facts["timestamped_text_messages"],
-        "attachment_count": input_facts["attachment_count"],
-        "thinking_messages": input_facts["thinking_messages"],
-        "tool_messages": input_facts["tool_messages"],
-        "branch_messages": input_facts["branch_messages"],
-    }
-
-
-def _mcp_detail_output_facts(payload: dict[str, Any]) -> dict[str, Any]:
-    messages = payload.get("messages")
-    if not isinstance(messages, list):
-        messages = []
-    role_counts: Counter[str] = Counter()
-    message_ids: list[str] = []
-    timestamped_messages = 0
-    for message in messages:
-        if not isinstance(message, dict):
-            continue
-        message_ids.append(str(message.get("id") or ""))
-        role_counts[_normalized_role_label(message.get("role"))] += 1
-        if message.get("timestamp"):
-            timestamped_messages += 1
-    return {
-        "conversation_id": str(payload.get("id") or ""),
-        "provider": str(payload.get("provider") or ""),
-        "title": payload.get("title"),
-        "created_at": payload.get("created_at"),
-        "updated_at": payload.get("updated_at"),
-        "messages": len(message_ids),
-        "message_ids": message_ids,
-        "role_counts": _sorted_counts(dict(role_counts)),
-        "timestamped_messages": timestamped_messages,
-    }
-
-
-def _json_like_output_facts(payload: dict[str, Any]) -> dict[str, Any]:
-    messages = payload.get("messages")
-    if not isinstance(messages, list):
-        messages = []
-    role_counts: Counter[str] = Counter()
-    message_ids: list[str] = []
-    timestamped_messages = 0
-    for message in messages:
-        if not isinstance(message, dict):
-            continue
-        message_ids.append(str(message.get("id", "")))
-        role_counts[_normalized_role_label(message.get("role"))] += 1
-        if message.get("timestamp"):
-            timestamped_messages += 1
-    return {
-        "conversation_id": str(payload.get("id") or ""),
-        "provider": str(payload.get("provider") or ""),
-        "title": payload.get("title"),
-        "date": payload.get("date"),
-        "messages": len(message_ids),
-        "message_ids": message_ids,
-        "role_counts": _sorted_counts(dict(role_counts)),
-        "timestamped_messages": timestamped_messages,
-    }
-
-
-def _csv_output_facts(csv_text: str) -> dict[str, Any]:
-    reader = csv.DictReader(io.StringIO(csv_text))
-    message_ids: list[str] = []
-    role_counts: Counter[str] = Counter()
-    timestamped_messages = 0
-    conversation_ids: Counter[str] = Counter()
-    for row in reader:
-        conversation_id = str(row.get("conversation_id") or "")
-        if conversation_id:
-            conversation_ids[conversation_id] += 1
-        message_ids.append(str(row.get("message_id") or ""))
-        role_counts[_normalized_role_label(row.get("role"))] += 1
-        if row.get("timestamp"):
-            timestamped_messages += 1
-    conversation_id = ""
-    if conversation_ids:
-        conversation_id = conversation_ids.most_common(1)[0][0]
-    return {
-        "conversation_id": conversation_id,
-        "messages": len(message_ids),
-        "message_ids": message_ids,
-        "role_counts": _sorted_counts(dict(role_counts)),
-        "timestamped_messages": timestamped_messages,
-        "provider": None,
-        "title": None,
-        "date": None,
-    }
-
-
-def _markdown_doc_output_facts(text: str) -> dict[str, Any]:
-    title: str | None = None
-    provider: str | None = None
-    has_date = False
-    message_sections = 0
-    role_counts: Counter[str] = Counter()
-    for line in text.splitlines():
-        if title is None and line.startswith("# "):
-            title = line[2:].strip()
-        elif line.startswith("**Provider**: "):
-            provider = line.split(": ", 1)[1].strip()
-        elif line.startswith("**Date**: "):
-            has_date = True
-        elif line.startswith("## "):
-            role = line[3:].strip()
-            message_sections += 1
-            role_counts[_normalized_role_label(role)] += 1
-    return {
-        "title": title,
-        "provider": provider,
-        "has_date": has_date,
-        "message_sections": message_sections,
-        "role_counts": _sorted_counts(dict(role_counts)),
-        "timestamp_lines": 0,
-        "branch_labels": 0,
-        "conversation_id": None,
-    }
-
-
-def _obsidian_output_facts(text: str) -> dict[str, Any]:
-    frontmatter: dict[str, Any] = {}
-    body = text
-    if text.startswith("---\n"):
-        parts = text.split("\n---\n", 1)
-        if len(parts) == 2:
-            _, body = parts
-            try:
-                import yaml
-
-                parsed = yaml.safe_load(parts[0].replace("---\n", "", 1))
-                if isinstance(parsed, dict):
-                    frontmatter = parsed
-            except Exception:
-                frontmatter = {}
-    body_facts = _markdown_doc_output_facts(body)
-    body_facts["conversation_id"] = str(frontmatter.get("id") or "")
-    body_facts["provider"] = str(frontmatter.get("provider") or "")
-    body_facts["has_date"] = frontmatter.get("date") is not None
-    return body_facts
-
-
-def _org_output_facts(text: str) -> dict[str, Any]:
-    title: str | None = None
-    provider: str | None = None
-    has_date = False
-    message_sections = 0
-    role_counts: Counter[str] = Counter()
-    for line in text.splitlines():
-        if line.startswith("#+TITLE: "):
-            title = line.split(": ", 1)[1].strip()
-        elif line.startswith("#+DATE: "):
-            has_date = True
-        elif line.startswith("#+PROPERTY: provider "):
-            provider = line.split("#+PROPERTY: provider ", 1)[1].strip()
-        elif line.startswith("* "):
-            role = line[2:].strip()
-            message_sections += 1
-            role_counts[_normalized_role_label(role)] += 1
-    return {
-        "title": title,
-        "provider": provider,
-        "has_date": has_date,
-        "message_sections": message_sections,
-        "role_counts": _sorted_counts(dict(role_counts)),
-        "timestamp_lines": 0,
-        "branch_labels": 0,
-        "conversation_id": None,
-    }
-
-
-def _html_output_facts(text: str) -> dict[str, Any]:
-    title_match = _HTML_TITLE_RE.search(text)
-    title = html_unescape(title_match.group(1).strip()) if title_match else None
-    badge_match = _HTML_BADGE_RE.search(text)
-    provider = badge_match.group(1).strip() if badge_match else None
-    meta_items = [html_unescape(item).strip() for item in _HTML_META_RE.findall(text)]
-    has_date = len(meta_items) >= 2
-    role_counts: Counter[str] = Counter(
-        _normalized_role_label(role.strip()) for role in _HTML_ROLE_RE.findall(text)
-    )
-    return {
-        "title": title,
-        "provider": provider,
-        "has_date": has_date,
-        "message_sections": sum(role_counts.values()),
-        "role_counts": _sorted_counts(dict(role_counts)),
-        "timestamp_lines": len(_HTML_TIMESTAMP_RE.findall(text)),
-        "branch_labels": len(_HTML_BRANCH_RE.findall(text)),
-        "conversation_id": None,
-    }
-
-
-@dataclass(frozen=True)
-class SemanticMetricCheck:
-    """One measurable preservation or declared-loss claim."""
-
-    metric: str
-    status: str
-    policy: str
-    input_value: Any
-    output_value: Any
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "metric": self.metric,
-            "status": self.status,
-            "policy": self.policy,
-            "input_value": self.input_value,
-            "output_value": self.output_value,
-        }
-
-
-@dataclass(frozen=True)
-class SemanticConversationProof:
-    """Semantic proof result for one rendered/exported conversation surface."""
-
-    conversation_id: str
-    provider: str
-    surface: str
-    input_facts: dict[str, Any]
-    output_facts: dict[str, Any]
-    checks: list[SemanticMetricCheck] = field(default_factory=list)
-
-    @property
-    def critical_loss_checks(self) -> list[SemanticMetricCheck]:
-        return [check for check in self.checks if check.status == "critical_loss"]
-
-    @property
-    def declared_loss_checks(self) -> list[SemanticMetricCheck]:
-        return [check for check in self.checks if check.status == "declared_loss"]
-
-    @property
-    def preserved_checks(self) -> list[SemanticMetricCheck]:
-        return [check for check in self.checks if check.status == "preserved"]
-
-    @property
-    def metric_summary(self) -> dict[str, dict[str, int]]:
-        summary: dict[str, dict[str, int]] = {}
-        for check in self.checks:
-            counts = summary.setdefault(check.metric, _empty_metric_counts())
-            counts[check.status] += 1
-        return dict(sorted(summary.items()))
-
-    @property
-    def is_clean(self) -> bool:
-        return not self.critical_loss_checks
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "conversation_id": self.conversation_id,
-            "provider": self.provider,
-            "surface": self.surface,
-            "input_facts": self.input_facts,
-            "output_facts": self.output_facts,
-            "summary": {
-                "preserved_checks": len(self.preserved_checks),
-                "declared_loss_checks": len(self.declared_loss_checks),
-                "critical_loss_checks": len(self.critical_loss_checks),
-                "metric_summary": self.metric_summary,
-                "clean": self.is_clean,
-            },
-            "checks": [check.to_dict() for check in self.checks],
-        }
-
-
-@dataclass(frozen=True)
-class ProviderSemanticProof:
-    """Per-provider semantic-preservation proof summary."""
-
-    provider: str
-    total_conversations: int = 0
-    clean_conversations: int = 0
-    critical_conversations: int = 0
-    preserved_checks: int = 0
-    declared_loss_checks: int = 0
-    critical_loss_checks: int = 0
-    metric_summary: dict[str, dict[str, int]] = field(default_factory=dict)
-
-    @property
-    def is_clean(self) -> bool:
-        return self.critical_conversations == 0
-
-    @property
-    def clean(self) -> bool:
-        return self.is_clean
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "provider": self.provider,
-            "total_conversations": self.total_conversations,
-            "clean_conversations": self.clean_conversations,
-            "critical_conversations": self.critical_conversations,
-            "preserved_checks": self.preserved_checks,
-            "declared_loss_checks": self.declared_loss_checks,
-            "critical_loss_checks": self.critical_loss_checks,
-            "metric_summary": dict(sorted(self.metric_summary.items())),
-            "clean": self.clean,
-        }
-
-
-def _build_provider_reports(
-    conversations: list[SemanticConversationProof],
-) -> dict[str, ProviderSemanticProof]:
-    provider_totals: dict[str, dict[str, Any]] = {}
-    for proof in conversations:
-        state = provider_totals.setdefault(
-            proof.provider,
-            {
-                "total_conversations": 0,
-                "clean_conversations": 0,
-                "critical_conversations": 0,
-                "preserved_checks": 0,
-                "declared_loss_checks": 0,
-                "critical_loss_checks": 0,
-                "metric_summary": {},
-            },
-        )
-        state["total_conversations"] += 1
-        if proof.is_clean:
-            state["clean_conversations"] += 1
-        else:
-            state["critical_conversations"] += 1
-        state["preserved_checks"] += len(proof.preserved_checks)
-        state["declared_loss_checks"] += len(proof.declared_loss_checks)
-        state["critical_loss_checks"] += len(proof.critical_loss_checks)
-        for metric, counts in proof.metric_summary.items():
-            metric_counts = state["metric_summary"].setdefault(metric, _empty_metric_counts())
-            for status, count in counts.items():
-                metric_counts[status] += count
-
-    return {
-        provider: ProviderSemanticProof(
-            provider=provider,
-            total_conversations=state["total_conversations"],
-            clean_conversations=state["clean_conversations"],
-            critical_conversations=state["critical_conversations"],
-            preserved_checks=state["preserved_checks"],
-            declared_loss_checks=state["declared_loss_checks"],
-            critical_loss_checks=state["critical_loss_checks"],
-            metric_summary=dict(sorted(state["metric_summary"].items())),
-        )
-        for provider, state in sorted(provider_totals.items())
-    }
-
-
-@dataclass(frozen=True)
-class SemanticProofReport:
-    """Aggregate semantic proof report for one output surface."""
-
-    surface: str
-    conversations: list[SemanticConversationProof]
-    provider_reports: dict[str, ProviderSemanticProof]
-    record_limit: int | None = None
-    record_offset: int = 0
-    provider_filters: list[str] = field(default_factory=list)
-
-    @property
-    def total_conversations(self) -> int:
-        return len(self.conversations)
-
-    @property
-    def provider_count(self) -> int:
-        return len(self.provider_reports)
-
-    @property
-    def providers(self) -> dict[str, ProviderSemanticProof]:
-        return self.provider_reports
-
-    @property
-    def clean_conversations(self) -> int:
-        return sum(1 for proof in self.conversations if proof.is_clean)
-
-    @property
-    def critical_conversations(self) -> int:
-        return sum(1 for proof in self.conversations if not proof.is_clean)
-
-    @property
-    def preserved_checks(self) -> int:
-        return sum(len(proof.preserved_checks) for proof in self.conversations)
-
-    @property
-    def declared_loss_checks(self) -> int:
-        return sum(len(proof.declared_loss_checks) for proof in self.conversations)
-
-    @property
-    def critical_loss_checks(self) -> int:
-        return sum(len(proof.critical_loss_checks) for proof in self.conversations)
-
-    @property
-    def metric_summary(self) -> dict[str, dict[str, int]]:
-        summary: dict[str, dict[str, int]] = {}
-        for proof in self.conversations:
-            for check in proof.checks:
-                metric = summary.setdefault(check.metric, _empty_metric_counts())
-                metric[check.status] += 1
-        return dict(sorted(summary.items()))
-
-    @property
-    def is_clean(self) -> bool:
-        return self.critical_conversations == 0
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "surface": self.surface,
-            "record_limit": self.record_limit if self.record_limit is not None else "all",
-            "record_offset": self.record_offset,
-            "provider_filters": list(self.provider_filters),
-            "summary": {
-                "total_conversations": self.total_conversations,
-                "provider_count": self.provider_count,
-                "clean_conversations": self.clean_conversations,
-                "critical_conversations": self.critical_conversations,
-                "preserved_checks": self.preserved_checks,
-                "declared_loss_checks": self.declared_loss_checks,
-                "critical_loss_checks": self.critical_loss_checks,
-                "metric_summary": self.metric_summary,
-                "clean": self.is_clean,
-            },
-            "providers": {
-                provider: stats.to_dict() for provider, stats in sorted(self.provider_reports.items())
-            },
-            "conversations": [proof.to_dict() for proof in self.conversations],
-        }
-
-
-@dataclass(frozen=True)
-class SemanticProofSuiteReport:
-    """Aggregate semantic proof report spanning multiple output surfaces."""
-
-    surface_reports: dict[str, SemanticProofReport]
-    record_limit: int | None = None
-    record_offset: int = 0
-    provider_filters: list[str] = field(default_factory=list)
-    surface_filters: list[str] = field(default_factory=list)
-
-    @property
-    def surfaces(self) -> dict[str, SemanticProofReport]:
-        return self.surface_reports
-
-    @property
-    def surface_count(self) -> int:
-        return len(self.surface_reports)
-
-    @property
-    def clean_surfaces(self) -> int:
-        return sum(1 for report in self.surface_reports.values() if report.is_clean)
-
-    @property
-    def critical_surfaces(self) -> int:
-        return sum(1 for report in self.surface_reports.values() if not report.is_clean)
-
-    @property
-    def total_conversations(self) -> int:
-        return sum(report.total_conversations for report in self.surface_reports.values())
-
-    @property
-    def preserved_checks(self) -> int:
-        return sum(report.preserved_checks for report in self.surface_reports.values())
-
-    @property
-    def declared_loss_checks(self) -> int:
-        return sum(report.declared_loss_checks for report in self.surface_reports.values())
-
-    @property
-    def critical_loss_checks(self) -> int:
-        return sum(report.critical_loss_checks for report in self.surface_reports.values())
-
-    @property
-    def metric_summary(self) -> dict[str, dict[str, int]]:
-        summary: dict[str, dict[str, int]] = {}
-        for report in self.surface_reports.values():
-            for metric, counts in report.metric_summary.items():
-                metric_counts = summary.setdefault(metric, _empty_metric_counts())
-                for status, count in counts.items():
-                    metric_counts[status] += count
-        return dict(sorted(summary.items()))
-
-    @property
-    def is_clean(self) -> bool:
-        return self.critical_surfaces == 0
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "record_limit": self.record_limit if self.record_limit is not None else "all",
-            "record_offset": self.record_offset,
-            "provider_filters": list(self.provider_filters),
-            "surface_filters": list(self.surface_filters),
-            "summary": {
-                "surface_count": self.surface_count,
-                "clean_surfaces": self.clean_surfaces,
-                "critical_surfaces": self.critical_surfaces,
-                "total_conversations": self.total_conversations,
-                "preserved_checks": self.preserved_checks,
-                "declared_loss_checks": self.declared_loss_checks,
-                "critical_loss_checks": self.critical_loss_checks,
-                "metric_summary": self.metric_summary,
-                "clean": self.is_clean,
-            },
-            "surfaces": {
-                surface: report.to_dict()
-                for surface, report in sorted(self.surface_reports.items())
-            },
-        }
-
 
 def resolve_semantic_surfaces(surfaces: list[str] | tuple[str, ...] | None) -> list[str]:
     """Normalize semantic-proof surface filters to canonical surface names."""
@@ -1037,1227 +158,6 @@ def resolve_semantic_surfaces(surfaces: list[str] | tuple[str, ...] | None) -> l
                 seen.add(alias)
                 resolved.append(alias)
     return resolved
-
-
-def prove_markdown_projection_semantics(
-    projection: ConversationRenderProjection,
-    markdown_text: str,
-) -> SemanticConversationProof:
-    """Compare a repository render projection to canonical markdown output."""
-    input_facts = _input_facts(projection)
-    output_facts = _canonical_markdown_output_facts(markdown_text)
-
-    checks = [
-        _critical_or_preserved(
-            metric="renderable_messages",
-            policy="canonical markdown must preserve every renderable message section",
-            input_value=input_facts["renderable_messages"],
-            output_value=output_facts["message_sections"],
-        ),
-        _critical_or_preserved(
-            metric="attachment_lines",
-            policy="canonical markdown must preserve attachment presence as attachment lines",
-            input_value=input_facts["attachment_count"],
-            output_value=output_facts["attachment_lines"],
-        ),
-        _critical_or_preserved(
-            metric="timestamp_lines",
-            policy="canonical markdown must preserve timestamps for renderable messages that have them",
-            input_value=input_facts["timestamped_renderable_messages"],
-            output_value=output_facts["timestamp_lines"],
-        ),
-        _critical_or_preserved(
-            metric="role_sections",
-            policy="canonical markdown must preserve renderable message role sections",
-            input_value=input_facts["renderable_role_counts"],
-            output_value=output_facts["role_section_counts"],
-        ),
-        _declared_loss_or_preserved(
-            metric="empty_messages",
-            policy="canonical markdown intentionally omits messages with no text and no attachments",
-            input_value=input_facts["empty_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="thinking_semantics",
-            policy="canonical markdown preserves display text but not typed thinking markers",
-            input_value=input_facts["thinking_messages"],
-            output_value=output_facts["typed_thinking_markers"],
-        ),
-        _declared_loss_or_preserved(
-            metric="tool_semantics",
-            policy="canonical markdown preserves display text but not typed tool markers",
-            input_value=input_facts["tool_messages"],
-            output_value=output_facts["typed_tool_markers"],
-        ),
-    ]
-
-    return SemanticConversationProof(
-        conversation_id=projection.conversation.conversation_id,
-        provider=projection.conversation.provider_name or "unknown",
-        surface="canonical_markdown_v1",
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
-
-
-def _prove_export_json_like_surface(
-    *,
-    conversation: Conversation,
-    rendered_text: str,
-    surface: str,
-) -> SemanticConversationProof:
-    try:
-        if surface == "export_yaml_v1":
-            import yaml
-
-            parsed = yaml.safe_load(rendered_text)
-        else:
-            parsed = json.loads(rendered_text)
-    except Exception:
-        parsed = {}
-    payload = parsed if isinstance(parsed, dict) else {}
-    input_facts = _conversation_input_facts(conversation)
-    output_facts = _json_like_output_facts(payload)
-    checks = [
-        _critical_or_preserved(
-            metric="conversation_id",
-            policy=f"{surface} must preserve the conversation identifier",
-            input_value=input_facts["conversation_id"],
-            output_value=output_facts["conversation_id"],
-        ),
-        _critical_or_preserved(
-            metric="provider_identity",
-            policy=f"{surface} must preserve provider identity",
-            input_value=input_facts["provider"],
-            output_value=output_facts["provider"],
-        ),
-        _critical_or_preserved(
-            metric="title_metadata",
-            policy=f"{surface} must preserve the display title",
-            input_value=input_facts["title"],
-            output_value=output_facts["title"],
-        ),
-        _critical_or_preserved(
-            metric="date_metadata",
-            policy=f"{surface} must preserve the display date value when present",
-            input_value=input_facts["date"],
-            output_value=output_facts["date"],
-        ),
-        _critical_or_preserved(
-            metric="message_entries",
-            policy=f"{surface} must preserve every message entry",
-            input_value=input_facts["total_messages"],
-            output_value=output_facts["messages"],
-        ),
-        _critical_or_preserved(
-            metric="message_ids",
-            policy=f"{surface} must preserve message identifiers",
-            input_value=input_facts["message_ids"],
-            output_value=output_facts["message_ids"],
-        ),
-        _critical_or_preserved(
-            metric="role_entries",
-            policy=f"{surface} must preserve message role distribution",
-            input_value=input_facts["text_role_counts"],
-            output_value=output_facts["role_counts"],
-        ),
-        _critical_or_preserved(
-            metric="timestamp_values",
-            policy=f"{surface} must preserve message timestamps",
-            input_value=input_facts["timestamped_text_messages"],
-            output_value=output_facts["timestamped_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="attachment_semantics",
-            policy=f"{surface} intentionally omits attachment payload semantics",
-            input_value=input_facts["attachment_count"],
-        ),
-        _declared_loss_or_preserved(
-            metric="thinking_semantics",
-            policy=f"{surface} preserves display text but not typed thinking markers",
-            input_value=input_facts["thinking_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="tool_semantics",
-            policy=f"{surface} preserves display text but not typed tool markers",
-            input_value=input_facts["tool_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="branch_structure",
-            policy=f"{surface} intentionally omits explicit branch topology",
-            input_value=input_facts["branch_messages"],
-        ),
-    ]
-    return SemanticConversationProof(
-        conversation_id=input_facts["conversation_id"],
-        provider=input_facts["provider"],
-        surface=surface,
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
-
-
-def _prove_export_csv_surface(
-    *,
-    conversation: Conversation,
-    rendered_text: str,
-) -> SemanticConversationProof:
-    input_facts = _conversation_input_facts(conversation)
-    output_facts = _csv_output_facts(rendered_text)
-    checks = [
-        _critical_or_preserved(
-            metric="conversation_id",
-            policy="export_csv_v1 must preserve the conversation identifier per row",
-            input_value=input_facts["conversation_id"],
-            output_value=output_facts["conversation_id"],
-        ),
-        _critical_or_preserved(
-            metric="text_messages",
-            policy="export_csv_v1 must preserve one row per text-bearing message",
-            input_value=input_facts["text_messages"],
-            output_value=output_facts["messages"],
-        ),
-        _critical_or_preserved(
-            metric="text_message_ids",
-            policy="export_csv_v1 must preserve identifiers for text-bearing messages",
-            input_value=input_facts["text_message_ids"],
-            output_value=output_facts["message_ids"],
-        ),
-        _critical_or_preserved(
-            metric="role_entries",
-            policy="export_csv_v1 must preserve roles for text-bearing messages",
-            input_value=input_facts["text_role_counts"],
-            output_value=output_facts["role_counts"],
-        ),
-        _critical_or_preserved(
-            metric="timestamp_values",
-            policy="export_csv_v1 must preserve timestamps for text-bearing messages",
-            input_value=input_facts["timestamped_text_messages"],
-            output_value=output_facts["timestamped_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="provider_identity",
-            policy="export_csv_v1 intentionally omits conversation-level provider metadata",
-            input_value=1 if input_facts["provider"] else 0,
-            output_value=0,
-        ),
-        _declared_loss_or_preserved(
-            metric="title_metadata",
-            policy="export_csv_v1 intentionally omits conversation-level title metadata",
-            input_value=1 if input_facts["title"] else 0,
-            output_value=0,
-        ),
-        _declared_loss_or_preserved(
-            metric="date_metadata",
-            policy="export_csv_v1 intentionally omits conversation-level date metadata",
-            input_value=1 if input_facts["date"] else 0,
-            output_value=0,
-        ),
-        _declared_loss_or_preserved(
-            metric="attachment_semantics",
-            policy="export_csv_v1 intentionally omits attachment payload semantics",
-            input_value=input_facts["attachment_count"],
-        ),
-        _declared_loss_or_preserved(
-            metric="thinking_semantics",
-            policy="export_csv_v1 preserves display text but not typed thinking markers",
-            input_value=input_facts["thinking_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="tool_semantics",
-            policy="export_csv_v1 preserves display text but not typed tool markers",
-            input_value=input_facts["tool_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="branch_structure",
-            policy="export_csv_v1 intentionally omits explicit branch topology",
-            input_value=input_facts["branch_messages"],
-        ),
-    ]
-    return SemanticConversationProof(
-        conversation_id=input_facts["conversation_id"],
-        provider=input_facts["provider"],
-        surface="export_csv_v1",
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
-
-
-def _prove_export_markdown_surface(
-    *,
-    conversation: Conversation,
-    rendered_text: str,
-) -> SemanticConversationProof:
-    input_facts = _conversation_input_facts(conversation)
-    output_facts = _markdown_doc_output_facts(rendered_text)
-    checks = [
-        _critical_or_preserved(
-            metric="title_metadata",
-            policy="export_markdown_v1 must preserve the display title",
-            input_value=input_facts["title"],
-            output_value=output_facts["title"],
-        ),
-        _critical_or_preserved(
-            metric="provider_identity",
-            policy="export_markdown_v1 must preserve provider identity at document level",
-            input_value=input_facts["provider"],
-            output_value=output_facts["provider"],
-        ),
-        _presence_check(
-            metric="date_metadata",
-            policy="export_markdown_v1 must preserve conversation date presence at document level",
-            input_value=input_facts["date"],
-            output_present=bool(output_facts["has_date"]),
-        ),
-        _critical_or_preserved(
-            metric="text_messages",
-            policy="export_markdown_v1 must preserve one section per text-bearing message",
-            input_value=input_facts["text_messages"],
-            output_value=output_facts["message_sections"],
-        ),
-        _critical_or_preserved(
-            metric="role_sections",
-            policy="export_markdown_v1 must preserve role sections for text-bearing messages",
-            input_value=input_facts["text_role_counts"],
-            output_value=output_facts["role_counts"],
-        ),
-        _declared_loss_or_preserved(
-            metric="timestamp_values",
-            policy="export_markdown_v1 intentionally omits per-message timestamps",
-            input_value=input_facts["timestamped_text_messages"],
-            output_value=output_facts["timestamp_lines"],
-        ),
-        _declared_loss_or_preserved(
-            metric="attachment_semantics",
-            policy="export_markdown_v1 intentionally omits attachment payload semantics",
-            input_value=input_facts["attachment_count"],
-        ),
-        _declared_loss_or_preserved(
-            metric="thinking_semantics",
-            policy="export_markdown_v1 preserves display text but not typed thinking markers",
-            input_value=input_facts["thinking_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="tool_semantics",
-            policy="export_markdown_v1 preserves display text but not typed tool markers",
-            input_value=input_facts["tool_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="branch_structure",
-            policy="export_markdown_v1 intentionally omits explicit branch topology",
-            input_value=input_facts["branch_messages"],
-            output_value=output_facts["branch_labels"],
-        ),
-    ]
-    return SemanticConversationProof(
-        conversation_id=input_facts["conversation_id"],
-        provider=input_facts["provider"],
-        surface="export_markdown_v1",
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
-
-
-def _prove_export_obsidian_surface(
-    *,
-    conversation: Conversation,
-    rendered_text: str,
-) -> SemanticConversationProof:
-    input_facts = _conversation_input_facts(conversation)
-    output_facts = _obsidian_output_facts(rendered_text)
-    checks = [
-        _critical_or_preserved(
-            metric="conversation_id",
-            policy="export_obsidian_v1 must preserve conversation identity in frontmatter",
-            input_value=input_facts["conversation_id"],
-            output_value=output_facts["conversation_id"],
-        ),
-        _critical_or_preserved(
-            metric="title_metadata",
-            policy="export_obsidian_v1 must preserve the display title",
-            input_value=input_facts["title"],
-            output_value=output_facts["title"],
-        ),
-        _critical_or_preserved(
-            metric="provider_identity",
-            policy="export_obsidian_v1 must preserve provider identity in frontmatter",
-            input_value=input_facts["provider"],
-            output_value=output_facts["provider"],
-        ),
-        _presence_check(
-            metric="date_metadata",
-            policy="export_obsidian_v1 must preserve conversation date presence in frontmatter",
-            input_value=input_facts["date"],
-            output_present=bool(output_facts["has_date"]),
-        ),
-        _critical_or_preserved(
-            metric="text_messages",
-            policy="export_obsidian_v1 must preserve one section per text-bearing message",
-            input_value=input_facts["text_messages"],
-            output_value=output_facts["message_sections"],
-        ),
-        _critical_or_preserved(
-            metric="role_sections",
-            policy="export_obsidian_v1 must preserve role sections for text-bearing messages",
-            input_value=input_facts["text_role_counts"],
-            output_value=output_facts["role_counts"],
-        ),
-        _declared_loss_or_preserved(
-            metric="timestamp_values",
-            policy="export_obsidian_v1 intentionally omits per-message timestamps",
-            input_value=input_facts["timestamped_text_messages"],
-            output_value=output_facts["timestamp_lines"],
-        ),
-        _declared_loss_or_preserved(
-            metric="attachment_semantics",
-            policy="export_obsidian_v1 intentionally omits attachment payload semantics",
-            input_value=input_facts["attachment_count"],
-        ),
-        _declared_loss_or_preserved(
-            metric="thinking_semantics",
-            policy="export_obsidian_v1 preserves display text but not typed thinking markers",
-            input_value=input_facts["thinking_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="tool_semantics",
-            policy="export_obsidian_v1 preserves display text but not typed tool markers",
-            input_value=input_facts["tool_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="branch_structure",
-            policy="export_obsidian_v1 intentionally omits explicit branch topology",
-            input_value=input_facts["branch_messages"],
-            output_value=output_facts["branch_labels"],
-        ),
-    ]
-    return SemanticConversationProof(
-        conversation_id=input_facts["conversation_id"],
-        provider=input_facts["provider"],
-        surface="export_obsidian_v1",
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
-
-
-def _prove_export_org_surface(
-    *,
-    conversation: Conversation,
-    rendered_text: str,
-) -> SemanticConversationProof:
-    input_facts = _conversation_input_facts(conversation)
-    output_facts = _org_output_facts(rendered_text)
-    checks = [
-        _critical_or_preserved(
-            metric="title_metadata",
-            policy="export_org_v1 must preserve the display title",
-            input_value=input_facts["title"],
-            output_value=output_facts["title"],
-        ),
-        _critical_or_preserved(
-            metric="provider_identity",
-            policy="export_org_v1 must preserve provider identity at document level",
-            input_value=input_facts["provider"],
-            output_value=output_facts["provider"],
-        ),
-        _presence_check(
-            metric="date_metadata",
-            policy="export_org_v1 must preserve conversation date presence at document level",
-            input_value=input_facts["date"],
-            output_present=bool(output_facts["has_date"]),
-        ),
-        _critical_or_preserved(
-            metric="text_messages",
-            policy="export_org_v1 must preserve one heading per text-bearing message",
-            input_value=input_facts["text_messages"],
-            output_value=output_facts["message_sections"],
-        ),
-        _critical_or_preserved(
-            metric="role_sections",
-            policy="export_org_v1 must preserve role headings for text-bearing messages",
-            input_value=input_facts["text_role_counts"],
-            output_value=output_facts["role_counts"],
-        ),
-        _declared_loss_or_preserved(
-            metric="timestamp_values",
-            policy="export_org_v1 intentionally omits per-message timestamps",
-            input_value=input_facts["timestamped_text_messages"],
-            output_value=output_facts["timestamp_lines"],
-        ),
-        _declared_loss_or_preserved(
-            metric="attachment_semantics",
-            policy="export_org_v1 intentionally omits attachment payload semantics",
-            input_value=input_facts["attachment_count"],
-        ),
-        _declared_loss_or_preserved(
-            metric="thinking_semantics",
-            policy="export_org_v1 preserves display text but not typed thinking markers",
-            input_value=input_facts["thinking_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="tool_semantics",
-            policy="export_org_v1 preserves display text but not typed tool markers",
-            input_value=input_facts["tool_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="branch_structure",
-            policy="export_org_v1 intentionally omits explicit branch topology",
-            input_value=input_facts["branch_messages"],
-            output_value=output_facts["branch_labels"],
-        ),
-    ]
-    return SemanticConversationProof(
-        conversation_id=input_facts["conversation_id"],
-        provider=input_facts["provider"],
-        surface="export_org_v1",
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
-
-
-def _prove_export_html_surface(
-    *,
-    conversation: Conversation,
-    rendered_text: str,
-) -> SemanticConversationProof:
-    input_facts = _conversation_input_facts(conversation)
-    output_facts = _html_output_facts(rendered_text)
-    checks = [
-        _critical_or_preserved(
-            metric="title_metadata",
-            policy="export_html_v1 must preserve the display title",
-            input_value=input_facts["title"],
-            output_value=output_facts["title"],
-        ),
-        _critical_or_preserved(
-            metric="provider_identity",
-            policy="export_html_v1 must preserve provider identity at document level",
-            input_value=input_facts["provider"],
-            output_value=output_facts["provider"],
-        ),
-        _presence_check(
-            metric="date_metadata",
-            policy="export_html_v1 must preserve conversation date presence at document level",
-            input_value=input_facts["date"],
-            output_present=bool(output_facts["has_date"]),
-        ),
-        _critical_or_preserved(
-            metric="text_messages",
-            policy="export_html_v1 must preserve visible message sections for text-bearing messages",
-            input_value=input_facts["text_messages"],
-            output_value=output_facts["message_sections"],
-        ),
-        _critical_or_preserved(
-            metric="role_sections",
-            policy="export_html_v1 must preserve visible role labels for text-bearing messages",
-            input_value=input_facts["text_role_counts"],
-            output_value=output_facts["role_counts"],
-        ),
-        _critical_or_preserved(
-            metric="timestamp_values",
-            policy="export_html_v1 must preserve visible message timestamps",
-            input_value=input_facts["timestamped_text_messages"],
-            output_value=output_facts["timestamp_lines"],
-        ),
-        _critical_or_preserved(
-            metric="branch_structure",
-            policy="export_html_v1 must preserve visible branch groupings for branched messages",
-            input_value=input_facts["branch_messages"],
-            output_value=output_facts["branch_labels"],
-        ),
-        _declared_loss_or_preserved(
-            metric="attachment_semantics",
-            policy="export_html_v1 intentionally omits attachment payload semantics",
-            input_value=input_facts["attachment_count"],
-        ),
-        _declared_loss_or_preserved(
-            metric="thinking_semantics",
-            policy="export_html_v1 preserves display text but not typed thinking markers",
-            input_value=input_facts["thinking_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="tool_semantics",
-            policy="export_html_v1 preserves display text but not typed tool markers",
-            input_value=input_facts["tool_messages"],
-        ),
-    ]
-    return SemanticConversationProof(
-        conversation_id=input_facts["conversation_id"],
-        provider=input_facts["provider"],
-        surface="export_html_v1",
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
-
-
-def prove_export_surface_semantics(
-    conversation: Conversation,
-    surface: str,
-    rendered_text: str,
-) -> SemanticConversationProof:
-    """Compare a conversation export surface to the canonical conversation semantics."""
-    if surface == "export_json_v1":
-        return _prove_export_json_like_surface(
-            conversation=conversation,
-            rendered_text=rendered_text,
-            surface=surface,
-        )
-    if surface == "export_yaml_v1":
-        return _prove_export_json_like_surface(
-            conversation=conversation,
-            rendered_text=rendered_text,
-            surface=surface,
-        )
-    if surface == "export_csv_v1":
-        return _prove_export_csv_surface(conversation=conversation, rendered_text=rendered_text)
-    if surface == "export_markdown_v1":
-        return _prove_export_markdown_surface(conversation=conversation, rendered_text=rendered_text)
-    if surface == "export_obsidian_v1":
-        return _prove_export_obsidian_surface(conversation=conversation, rendered_text=rendered_text)
-    if surface == "export_org_v1":
-        return _prove_export_org_surface(conversation=conversation, rendered_text=rendered_text)
-    if surface == "export_html_v1":
-        return _prove_export_html_surface(conversation=conversation, rendered_text=rendered_text)
-    raise ValueError(f"Unsupported semantic surface: {surface}")
-
-
-def _summary_record_from_conversation(conversation: Conversation) -> ConversationSummary:
-    return ConversationSummary(
-        id=conversation.id,
-        provider=conversation.provider,
-        title=conversation.title,
-        created_at=conversation.created_at,
-        updated_at=conversation.updated_at,
-        provider_meta=conversation.provider_meta,
-        metadata=conversation.metadata,
-        parent_id=conversation.parent_id,
-        branch_type=conversation.branch_type,
-        message_count=len(conversation.messages),
-    )
-
-
-def _prove_query_summary_json_like_surface(
-    *,
-    summary: ConversationSummary,
-    message_count: int,
-    rendered_text: str,
-    surface: str,
-) -> SemanticConversationProof:
-    try:
-        if surface == "query_summary_yaml_v1":
-            import yaml
-
-            parsed = yaml.safe_load(rendered_text)
-        else:
-            parsed = json.loads(rendered_text)
-    except Exception:
-        parsed = []
-
-    row = parsed[0] if isinstance(parsed, list) and parsed else {}
-    payload = row if isinstance(row, dict) else {}
-    input_facts = _summary_input_facts(summary, message_count)
-    output_facts = _summary_output_facts(payload)
-    checks = [
-        _critical_or_preserved(
-            metric="conversation_id",
-            policy=f"{surface} must preserve the conversation identifier",
-            input_value=input_facts["conversation_id"],
-            output_value=output_facts["conversation_id"],
-        ),
-        _critical_or_preserved(
-            metric="provider_identity",
-            policy=f"{surface} must preserve provider identity",
-            input_value=input_facts["provider"],
-            output_value=output_facts["provider"],
-        ),
-        _critical_or_preserved(
-            metric="title_metadata",
-            policy=f"{surface} must preserve the display title",
-            input_value=input_facts["title"],
-            output_value=output_facts["title"],
-        ),
-        _critical_or_preserved(
-            metric="date_metadata",
-            policy=f"{surface} must preserve the summary display date",
-            input_value=input_facts["date"],
-            output_value=output_facts["date"],
-        ),
-        _critical_or_preserved(
-            metric="message_count",
-            policy=f"{surface} must preserve summary message counts",
-            input_value=input_facts["messages"],
-            output_value=output_facts["messages"],
-        ),
-        _critical_or_preserved(
-            metric="tag_values",
-            policy=f"{surface} must preserve summary tag values",
-            input_value=input_facts["tags"],
-            output_value=output_facts["tags"],
-        ),
-        _critical_or_preserved(
-            metric="summary_text",
-            policy=f"{surface} must preserve summary text",
-            input_value=input_facts["summary"],
-            output_value=output_facts["summary"],
-        ),
-    ]
-    return SemanticConversationProof(
-        conversation_id=input_facts["conversation_id"],
-        provider=input_facts["provider"],
-        surface=surface,
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
-
-
-def _prove_query_summary_csv_surface(
-    *,
-    summary: ConversationSummary,
-    message_count: int,
-    rendered_text: str,
-) -> SemanticConversationProof:
-    rows = _summary_csv_output_facts(rendered_text)
-    output_facts = rows[0] if rows else {}
-    input_facts = _summary_input_facts(summary, message_count)
-    checks = [
-        _critical_or_preserved(
-            metric="conversation_id",
-            policy="query_summary_csv_v1 must preserve the conversation identifier",
-            input_value=input_facts["conversation_id"],
-            output_value=output_facts.get("conversation_id", ""),
-        ),
-        _critical_or_preserved(
-            metric="provider_identity",
-            policy="query_summary_csv_v1 must preserve provider identity",
-            input_value=input_facts["provider"],
-            output_value=output_facts.get("provider", ""),
-        ),
-        _critical_or_preserved(
-            metric="title_metadata",
-            policy="query_summary_csv_v1 must preserve the display title",
-            input_value=input_facts["title"],
-            output_value=output_facts.get("title"),
-        ),
-        _critical_or_preserved(
-            metric="date_metadata",
-            policy="query_summary_csv_v1 must preserve the summary display date",
-            input_value=input_facts["date"][:10] if input_facts["date"] else None,
-            output_value=output_facts.get("date"),
-        ),
-        _critical_or_preserved(
-            metric="message_count",
-            policy="query_summary_csv_v1 must preserve summary message counts",
-            input_value=input_facts["messages"],
-            output_value=output_facts.get("messages", 0),
-        ),
-        _critical_or_preserved(
-            metric="tag_values",
-            policy="query_summary_csv_v1 must preserve summary tag values",
-            input_value=input_facts["tags"],
-            output_value=output_facts.get("tags", []),
-        ),
-        _critical_or_preserved(
-            metric="summary_text",
-            policy="query_summary_csv_v1 must preserve summary text",
-            input_value=input_facts["summary"],
-            output_value=output_facts.get("summary"),
-        ),
-    ]
-    return SemanticConversationProof(
-        conversation_id=input_facts["conversation_id"],
-        provider=input_facts["provider"],
-        surface="query_summary_csv_v1",
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
-
-
-def _prove_query_summary_text_surface(
-    *,
-    summary: ConversationSummary,
-    message_count: int,
-    rendered_text: str,
-) -> SemanticConversationProof:
-    rows = _summary_text_output_facts(rendered_text)
-    output_facts = rows[0] if rows else {}
-    input_facts = _summary_input_facts(summary, message_count)
-    raw_title = input_facts["title"] or input_facts["conversation_id"][:20]
-    expected_title = (raw_title[:47] + "...") if len(raw_title) > 50 else raw_title
-    checks = [
-        _critical_or_preserved(
-            metric="conversation_id_prefix",
-            policy="query_summary_text_v1 must preserve the visible conversation id prefix",
-            input_value=input_facts["conversation_id"][:24],
-            output_value=output_facts.get("conversation_id_prefix", ""),
-        ),
-        _critical_or_preserved(
-            metric="provider_identity",
-            policy="query_summary_text_v1 must preserve provider identity",
-            input_value=input_facts["provider"],
-            output_value=output_facts.get("provider", ""),
-        ),
-        _critical_or_preserved(
-            metric="date_metadata",
-            policy="query_summary_text_v1 must preserve the visible summary date",
-            input_value=input_facts["date"][:10] if input_facts["date"] else None,
-            output_value=output_facts.get("date"),
-        ),
-        _critical_or_preserved(
-            metric="title_projection",
-            policy="query_summary_text_v1 must preserve the deterministic visible title projection",
-            input_value=expected_title,
-            output_value=output_facts.get("title", ""),
-        ),
-        _critical_or_preserved(
-            metric="message_count",
-            policy="query_summary_text_v1 must preserve summary message counts",
-            input_value=input_facts["messages"],
-            output_value=output_facts.get("messages", 0),
-        ),
-        _declared_loss_or_preserved(
-            metric="tag_values",
-            policy="query_summary_text_v1 intentionally omits tag values",
-            input_value=len(input_facts["tags"]),
-        ),
-        _declared_loss_or_preserved(
-            metric="summary_text",
-            policy="query_summary_text_v1 intentionally omits summary text",
-            input_value=1 if input_facts["summary"] else 0,
-        ),
-    ]
-    return SemanticConversationProof(
-        conversation_id=input_facts["conversation_id"],
-        provider=input_facts["provider"],
-        surface="query_summary_text_v1",
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
-
-
-def _prove_query_stream_plaintext_surface(
-    *,
-    conversation: Conversation,
-    rendered_text: str,
-) -> SemanticConversationProof:
-    input_facts = _stream_input_facts(conversation)
-    output_facts = _stream_plaintext_output_facts(rendered_text)
-    checks = [
-        _critical_or_preserved(
-            metric="text_messages",
-            policy="query_stream_plaintext_v1 must preserve one visible block per text-bearing message",
-            input_value=input_facts["text_messages"],
-            output_value=output_facts["message_sections"],
-        ),
-        _critical_or_preserved(
-            metric="role_sections",
-            policy="query_stream_plaintext_v1 must preserve visible role labels for streamed messages",
-            input_value=input_facts["text_role_counts"],
-            output_value=output_facts["role_counts"],
-        ),
-        _declared_loss_or_preserved(
-            metric="title_metadata",
-            policy="query_stream_plaintext_v1 intentionally omits title metadata",
-            input_value=1 if input_facts["title"] else 0,
-        ),
-        _declared_loss_or_preserved(
-            metric="provider_identity",
-            policy="query_stream_plaintext_v1 intentionally omits provider metadata",
-            input_value=1 if input_facts["provider"] else 0,
-        ),
-        _declared_loss_or_preserved(
-            metric="date_metadata",
-            policy="query_stream_plaintext_v1 intentionally omits date metadata",
-            input_value=1 if input_facts["date"] else 0,
-        ),
-        _declared_loss_or_preserved(
-            metric="timestamp_values",
-            policy="query_stream_plaintext_v1 intentionally omits per-message timestamps",
-            input_value=input_facts["timestamped_text_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="attachment_semantics",
-            policy="query_stream_plaintext_v1 intentionally omits attachment semantics",
-            input_value=input_facts["attachment_count"],
-        ),
-        _declared_loss_or_preserved(
-            metric="thinking_semantics",
-            policy="query_stream_plaintext_v1 preserves display text but not typed thinking markers",
-            input_value=input_facts["thinking_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="tool_semantics",
-            policy="query_stream_plaintext_v1 preserves display text but not typed tool markers",
-            input_value=input_facts["tool_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="branch_structure",
-            policy="query_stream_plaintext_v1 intentionally omits explicit branch topology",
-            input_value=input_facts["branch_messages"],
-        ),
-    ]
-    return SemanticConversationProof(
-        conversation_id=str(conversation.id),
-        provider=str(conversation.provider),
-        surface="query_stream_plaintext_v1",
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
-
-
-def _prove_query_stream_markdown_surface(
-    *,
-    conversation: Conversation,
-    rendered_text: str,
-) -> SemanticConversationProof:
-    input_facts = _stream_input_facts(conversation)
-    output_facts = _stream_markdown_output_facts(rendered_text)
-    checks = [
-        _critical_or_preserved(
-            metric="title_metadata",
-            policy="query_stream_markdown_v1 must preserve the conversation title",
-            input_value=input_facts["title"],
-            output_value=output_facts["title"],
-        ),
-        _critical_or_preserved(
-            metric="provider_identity",
-            policy="query_stream_markdown_v1 must preserve provider identity in the stream header",
-            input_value=input_facts["provider"],
-            output_value=output_facts["provider"],
-        ),
-        _presence_check(
-            metric="date_metadata",
-            policy="query_stream_markdown_v1 must preserve conversation date presence in the stream header",
-            input_value=input_facts["date"],
-            output_present=bool(output_facts["has_date"]),
-        ),
-        _critical_or_preserved(
-            metric="text_messages",
-            policy="query_stream_markdown_v1 must preserve one visible section per text-bearing message",
-            input_value=input_facts["text_messages"],
-            output_value=output_facts["message_sections"],
-        ),
-        _critical_or_preserved(
-            metric="role_sections",
-            policy="query_stream_markdown_v1 must preserve visible role headings for streamed messages",
-            input_value=input_facts["text_role_counts"],
-            output_value=output_facts["role_counts"],
-        ),
-        _critical_or_preserved(
-            metric="footer_count",
-            policy="query_stream_markdown_v1 must report the number of emitted messages honestly",
-            input_value=input_facts["text_messages"],
-            output_value=output_facts["footer_count"],
-        ),
-        _declared_loss_or_preserved(
-            metric="timestamp_values",
-            policy="query_stream_markdown_v1 intentionally omits per-message timestamps",
-            input_value=input_facts["timestamped_text_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="attachment_semantics",
-            policy="query_stream_markdown_v1 intentionally omits attachment semantics",
-            input_value=input_facts["attachment_count"],
-        ),
-        _declared_loss_or_preserved(
-            metric="thinking_semantics",
-            policy="query_stream_markdown_v1 preserves display text but not typed thinking markers",
-            input_value=input_facts["thinking_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="tool_semantics",
-            policy="query_stream_markdown_v1 preserves display text but not typed tool markers",
-            input_value=input_facts["tool_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="branch_structure",
-            policy="query_stream_markdown_v1 intentionally omits explicit branch topology",
-            input_value=input_facts["branch_messages"],
-        ),
-    ]
-    return SemanticConversationProof(
-        conversation_id=str(conversation.id),
-        provider=str(conversation.provider),
-        surface="query_stream_markdown_v1",
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
-
-
-def _prove_query_stream_json_lines_surface(
-    *,
-    conversation: Conversation,
-    rendered_text: str,
-) -> SemanticConversationProof:
-    input_facts = _stream_input_facts(conversation)
-    output_facts = _stream_json_lines_output_facts(rendered_text)
-    checks = [
-        _critical_or_preserved(
-            metric="conversation_id",
-            policy="query_stream_json_lines_v1 must preserve the conversation identifier in the header",
-            input_value=input_facts["conversation_id"],
-            output_value=output_facts["conversation_id"],
-        ),
-        _critical_or_preserved(
-            metric="title_metadata",
-            policy="query_stream_json_lines_v1 must preserve the title in the header",
-            input_value=input_facts["title"],
-            output_value=output_facts["title"],
-        ),
-        _critical_or_preserved(
-            metric="provider_identity",
-            policy="query_stream_json_lines_v1 must preserve provider identity in the header",
-            input_value=input_facts["provider"],
-            output_value=output_facts["provider"],
-        ),
-        _critical_or_preserved(
-            metric="date_metadata",
-            policy="query_stream_json_lines_v1 must preserve the conversation date in the header",
-            input_value=input_facts["date"],
-            output_value=output_facts["date"],
-        ),
-        _critical_or_preserved(
-            metric="text_message_ids",
-            policy="query_stream_json_lines_v1 must preserve identifiers for emitted messages",
-            input_value=input_facts["text_message_ids"],
-            output_value=output_facts["message_ids"],
-        ),
-        _critical_or_preserved(
-            metric="role_sections",
-            policy="query_stream_json_lines_v1 must preserve role distribution for emitted messages",
-            input_value=input_facts["text_role_counts"],
-            output_value=output_facts["role_counts"],
-        ),
-        _critical_or_preserved(
-            metric="timestamp_values",
-            policy="query_stream_json_lines_v1 must preserve timestamps for emitted messages",
-            input_value=input_facts["timestamped_text_messages"],
-            output_value=output_facts["timestamped_messages"],
-        ),
-        _critical_or_preserved(
-            metric="footer_count",
-            policy="query_stream_json_lines_v1 must report the number of emitted messages honestly",
-            input_value=input_facts["text_messages"],
-            output_value=output_facts["footer_count"],
-        ),
-        _declared_loss_or_preserved(
-            metric="attachment_semantics",
-            policy="query_stream_json_lines_v1 intentionally omits attachment semantics",
-            input_value=input_facts["attachment_count"],
-        ),
-        _declared_loss_or_preserved(
-            metric="thinking_semantics",
-            policy="query_stream_json_lines_v1 preserves display text but not typed thinking markers",
-            input_value=input_facts["thinking_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="tool_semantics",
-            policy="query_stream_json_lines_v1 preserves display text but not typed tool markers",
-            input_value=input_facts["tool_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="branch_structure",
-            policy="query_stream_json_lines_v1 intentionally omits explicit branch topology",
-            input_value=input_facts["branch_messages"],
-        ),
-    ]
-    return SemanticConversationProof(
-        conversation_id=str(conversation.id),
-        provider=str(conversation.provider),
-        surface="query_stream_json_lines_v1",
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
-
-
-def _prove_mcp_summary_surface(
-    *,
-    summary: ConversationSummary,
-    message_count: int,
-    rendered_text: str,
-) -> SemanticConversationProof:
-    try:
-        parsed = json.loads(rendered_text)
-    except Exception:
-        parsed = []
-    row = parsed[0] if isinstance(parsed, list) and parsed else {}
-    payload = row if isinstance(row, dict) else {}
-    input_facts = {
-        "conversation_id": str(summary.id),
-        "provider": str(summary.provider),
-        "title": summary.display_title,
-        "messages": message_count,
-        "created_at": summary.created_at.isoformat() if summary.created_at else None,
-        "updated_at": summary.updated_at.isoformat() if summary.updated_at else None,
-        "tags": list(summary.tags),
-        "summary": summary.summary,
-    }
-    output_facts = _mcp_summary_output_facts(payload)
-    checks = [
-        _critical_or_preserved(
-            metric="conversation_id",
-            policy="mcp_summary_json_v1 must preserve the conversation identifier",
-            input_value=input_facts["conversation_id"],
-            output_value=output_facts["conversation_id"],
-        ),
-        _critical_or_preserved(
-            metric="provider_identity",
-            policy="mcp_summary_json_v1 must preserve provider identity",
-            input_value=input_facts["provider"],
-            output_value=output_facts["provider"],
-        ),
-        _critical_or_preserved(
-            metric="title_metadata",
-            policy="mcp_summary_json_v1 must preserve the display title",
-            input_value=input_facts["title"],
-            output_value=output_facts["title"],
-        ),
-        _critical_or_preserved(
-            metric="message_count",
-            policy="mcp_summary_json_v1 must preserve summary message counts",
-            input_value=input_facts["messages"],
-            output_value=output_facts["messages"],
-        ),
-        _critical_or_preserved(
-            metric="created_at",
-            policy="mcp_summary_json_v1 must preserve the created_at timestamp when present",
-            input_value=input_facts["created_at"],
-            output_value=output_facts["created_at"],
-        ),
-        _critical_or_preserved(
-            metric="updated_at",
-            policy="mcp_summary_json_v1 must preserve the updated_at timestamp when present",
-            input_value=input_facts["updated_at"],
-            output_value=output_facts["updated_at"],
-        ),
-        _declared_loss_or_preserved(
-            metric="tag_values",
-            policy="mcp_summary_json_v1 intentionally omits tags",
-            input_value=len(input_facts["tags"]),
-        ),
-        _declared_loss_or_preserved(
-            metric="summary_text",
-            policy="mcp_summary_json_v1 intentionally omits summary text",
-            input_value=1 if input_facts["summary"] else 0,
-        ),
-    ]
-    return SemanticConversationProof(
-        conversation_id=input_facts["conversation_id"],
-        provider=input_facts["provider"],
-        surface="mcp_summary_json_v1",
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
-
-
-def _prove_mcp_detail_surface(
-    *,
-    conversation: Conversation,
-    rendered_text: str,
-) -> SemanticConversationProof:
-    try:
-        parsed = json.loads(rendered_text)
-    except Exception:
-        parsed = {}
-    payload = parsed if isinstance(parsed, dict) else {}
-    input_facts = _mcp_detail_input_facts(conversation)
-    output_facts = _mcp_detail_output_facts(payload)
-    checks = [
-        _critical_or_preserved(
-            metric="conversation_id",
-            policy="mcp_detail_json_v1 must preserve the conversation identifier",
-            input_value=input_facts["conversation_id"],
-            output_value=output_facts["conversation_id"],
-        ),
-        _critical_or_preserved(
-            metric="provider_identity",
-            policy="mcp_detail_json_v1 must preserve provider identity",
-            input_value=input_facts["provider"],
-            output_value=output_facts["provider"],
-        ),
-        _critical_or_preserved(
-            metric="title_metadata",
-            policy="mcp_detail_json_v1 must preserve the display title",
-            input_value=input_facts["title"],
-            output_value=output_facts["title"],
-        ),
-        _critical_or_preserved(
-            metric="created_at",
-            policy="mcp_detail_json_v1 must preserve the created_at timestamp when present",
-            input_value=input_facts["created_at"],
-            output_value=output_facts["created_at"],
-        ),
-        _critical_or_preserved(
-            metric="updated_at",
-            policy="mcp_detail_json_v1 must preserve the updated_at timestamp when present",
-            input_value=input_facts["updated_at"],
-            output_value=output_facts["updated_at"],
-        ),
-        _critical_or_preserved(
-            metric="message_entries",
-            policy="mcp_detail_json_v1 must preserve every message entry",
-            input_value=input_facts["messages"],
-            output_value=output_facts["messages"],
-        ),
-        _critical_or_preserved(
-            metric="message_ids",
-            policy="mcp_detail_json_v1 must preserve message identifiers",
-            input_value=input_facts["message_ids"],
-            output_value=output_facts["message_ids"],
-        ),
-        _critical_or_preserved(
-            metric="role_entries",
-            policy="mcp_detail_json_v1 must preserve message role distribution",
-            input_value=input_facts["role_counts"],
-            output_value=output_facts["role_counts"],
-        ),
-        _critical_or_preserved(
-            metric="timestamp_values",
-            policy="mcp_detail_json_v1 must preserve message timestamps",
-            input_value=input_facts["timestamped_messages"],
-            output_value=output_facts["timestamped_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="attachment_semantics",
-            policy="mcp_detail_json_v1 intentionally omits attachment payload semantics",
-            input_value=input_facts["attachment_count"],
-        ),
-        _declared_loss_or_preserved(
-            metric="thinking_semantics",
-            policy="mcp_detail_json_v1 preserves display text but not typed thinking markers",
-            input_value=input_facts["thinking_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="tool_semantics",
-            policy="mcp_detail_json_v1 preserves display text but not typed tool markers",
-            input_value=input_facts["tool_messages"],
-        ),
-        _declared_loss_or_preserved(
-            metric="branch_structure",
-            policy="mcp_detail_json_v1 intentionally omits explicit branch topology",
-            input_value=input_facts["branch_messages"],
-        ),
-    ]
-    return SemanticConversationProof(
-        conversation_id=input_facts["conversation_id"],
-        provider=input_facts["provider"],
-        surface="mcp_detail_json_v1",
-        input_facts=input_facts,
-        output_facts=output_facts,
-        checks=checks,
-    )
 
 
 def _empty_surface_report(
@@ -2312,7 +212,10 @@ async def _prove_semantic_surface_suite_async(
             "query_summary_json_v1",
             "query_summary_yaml_v1",
         }
-        need_message_counts = any(surface in summary_surfaces or surface == "mcp_summary_json_v1" for surface in surfaces)
+        need_message_counts = any(
+            surface in summary_surfaces or surface == "mcp_summary_json_v1"
+            for surface in surfaces
+        )
         need_full_conversations = any(
             surface in export_surfaces
             or surface in stream_surfaces
